fix a small compiler warning in libicq200
basically it complained 'cause we were comparing
between unsigned shor and unsigned int, which was
quite easy to fix

+++ libicq2000/src/FileTransferClient.cpp
@@ -201,7 +201,7 @@ namespace ICQ2000
   {
     if (m_recv.empty()) return;
 
-    unsigned short length;
+    unsigned int length;
 
     while (!m_recv.empty())
     {
